# Refactor: Use `Object.prototype` directly