use `Object.prototype` directly, instead of `Object.getPrototypeOf({})`