@fisker 2020-01-07
@sindresorhus restored other changes, only replace `Object.getPrototypeOf({})` with `Object.prototype`