@fisker 2020-01-08
@sindresorhus Can you make a release?