@sindresorhus 2020-01-08
@fisker Why? Does it fix anything? I don't usually do new releases for refactorings and cosmetic changes.