@fisker 2020-01-08
@sindresorhus It's faster, no?