@sindresorhus 2020-01-11
Do you have any benchmark to prove that? And even if it's faster in a micro-benchmark, it most likely has no impact at all on real-world code. V8 is pretty good at optimizing stuff.