@fisker 2020-01-11
```js
const oldOne = value => {
	if (Object.prototype.toString.call(value) !== '[object Object]') {
		return false;
	}

	const prototype = Object.getPrototypeOf(value);
	return prototype === null || prototype === Object.getPrototypeOf({});
}

const newOne = value => {
	if (Object.prototype.toString.call(value) !== '[object Object]') {
		return false;
	}

	const prototype = Object.getPrototypeOf(value);
	return prototype === null || prototype === Object.prototype;
}

console.time('old')
for (let i = 0; i < 1e8; i++) {
  oldOne('')
  oldOne(Object.create(null))
  oldOne({})
}
console.timeEnd('old')

console.time('new')
for (let i = 0; i < 1e8; i++) {
  newOne('')
  newOne(Object.create(null))
  newOne({})
}
console.timeEnd('new')
```

About 20% faster

```text
$ nvm use 12.13.1
Now using node v12.13.1 (64-bit)

$ node 1.js
old: 9057.424ms
new: 7586.042ms

$ nvm use 10.17.0
Now using node v10.17.0 (64-bit)

$ node 1.js
old: 10626.064ms
new: 8113.567ms
```