@fisker 2020-01-11
If only test last branch `newOne({})`, 

```text
old: 4453.476ms
new: 2380.431ms
```

@sindresorhus