@sindresorhus 2020-01-21
I'll do a new release so we don't have to waste more time on this conversation (I mean that in the kindest way possible. I've had similar conversations like this all too many times), but I still strongly disagree it has any practical benefit. You're measuring a simple function in isolation. That is no proof how it will perform in real code.